% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serve}
\alias{serve}
\title{Serve a CRAN-like Package Repository}
\usage{
serve(repo, repo_name = "Cranium", host = "127.0.0.1", port = 8000,
  detach = FALSE, use_archive = TRUE, fields = NULL)
}
\arguments{
\item{repo}{The location of the package repository.}

\item{repo_name}{The name of the repository, written to \code{DESCRIPTION}
files.}

\item{host}{An IPv4 address owned by the server. Defaults to localhost.}

\item{port}{The port to run the server on.}

\item{detach}{Whether the server should run in the foreground or return
immediately and run in the background.}

\item{use_archive}{Whether old source packages should be moved to an Archive.}

\item{fields}{Metadata for each package to add to the index. When
\code{NULL}, use the defaults. See \code{\link[tools]{write_PACKAGES}}.}
}
\value{
When \code{detach = TRUE}, this function returns a handle that can be passed
to \code{\link{stopServer}}. Otherwise it will not return at all, and must
be interrupted from the console.
}
\description{
Starts a local web server to serve packages from the repository.
}
